requirements.txt
abei==0.0.3
Django==2.2.8
django-filter==2.4.0
django-rest-swagger==2.2.0
djangorestframework==3.10.3
drf-extensions==0.5.0
mb-drf-extensions==0.1.3
